+++ packages/react/src/components/gdpr-consent-dialog/constants.ts
@@ -1,0 +1,7 @@
+export const CONSENT_LEVELS_WITHOUT_NONE = [
+  "necessary",
+  "analytics",
+  "marketing",
+  "retargeting",
+  "tracking",
+] as const;
